ci: expose `CC_TEST_REPORTER_ID` (#143)
In [GitHub Action](https://docs.github.com/en/actions/reference/encrypted-secrets#using-encrypted-secrets-in-a-workflow), secrets are only exposed on `push` from a branch of the same repo. It makes it impossible to publish a coverage report.

As stated in the [codeclimate documentation](https://docs.codeclimate.com/docs/finding-your-test-coverage-token#should-i-keep-my-test-reporter-id-secret) `CC_TEST_REPORTER_ID` can't be considered as a secret.